+++ src/arc-validate/CLIArgs/ValidateArgs.fs
@@ -7,6 +7,8 @@ type ValidateArgs =
     | [<Unique; AltCommandLine("-p")>] Package of package_name:string
     | [<Unique; AltCommandLine("-v")>] Package_Version of package_version:string
     | [<Unique>] Specification_Version of specification_version: string
+    | [<Unique>] Source_Branch of branch: string
+    | [<Unique>] Source_Commit_Hash of commit_hash: string
 
     interface IArgParserTemplate with
         member s.Usage =
@@ -16,3 +18,5 @@ type ValidateArgs =
             | Package _       -> "Optional. Specify a validation package to use on top of the default validation for invenio export. Default: no package is used, only structural validation for invenio export."
             | Package_Version _ -> "Optional. Specify a version of the validation package to use. If no version is specified, the latest version will be used."
             | Specification_Version _ -> "Optional. Only has an effect if no package is specified via '-p' Specify a version of the ARC specification to validate against. Default: 'latest'."
+            | Source_Branch _ -> "Optional. Record the source branch in generated validation outputs."
+            | Source_Commit_Hash _ -> "Optional. Record the source commit hash in generated validation outputs."